add: mvn archetype, spring boot 2.7, rest api for merging docx files refactor: removed eclipse configs, moved concat logic to service
