--- .classpath
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-  <classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
-  <classpathentry kind="src" path="src"/>
-  <classpathentry kind="output" path="bin"/>
-  <classpathentry kind="lib" exported="true" path="lib/poi-ooxml-3.7-20101029.jar" />
-  <classpathentry kind="lib" exported="true" path="lib/poi-3.7-20101029.jar" />
-  <classpathentry kind="lib" exported="true" path="lib/ooxml-schemas-1.1.jar" />
-  <classpathentry kind="lib" exported="true" path="lib/xmlbeans-2.3.0.jar" />
-  <classpathentry kind="lib" exported="true" path="lib/dom4j-1.6.1.jar" />
-</classpath>
