--- .project
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<projectDescription>
-  <name>MergeDOCX</name>
-  <comment/>
-  <projects>&#xD;
-    </projects>
-  <buildSpec>
-    <buildCommand>
-      <name>org.eclipse.jdt.core.javabuilder</name>
-      <arguments>&#xD;
-            </arguments>
-    </buildCommand>
-  </buildSpec>
-  <natures>
-    <nature>org.eclipse.jdt.core.javanature</nature>
-  </natures>
-</projectDescription>
