+++ src/main/java/com/example/docxconcat/MergeDOCX.java
@@ -1,0 +1,13 @@
+package com.example.docxconcat;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MergeDOCX {
+
+    public static void main(String[] args) {
+        SpringApplication.run(MergeDOCX.class, args);
+    }
+
+}
