+++ src/main/java/com/example/docxconcat/configuration/ApplicationConfiguration.java
@@ -1,0 +1,52 @@
+package com.example.docxconcat.configuration;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.security.SecurityRequirement;
+import org.modelmapper.Conditions;
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.List;
+
+/**
+ * @author ogbozoyan
+ * @date 08.07.2023
+ */
+@Configuration
+public class ApplicationConfiguration {
+    @Bean(name = "patchingMapper")
+    ModelMapper patchingModelMapper() {
+        ModelMapper modelMapper = new ModelMapper();
+        modelMapper.getConfiguration()
+                .setAmbiguityIgnored(true)
+                .setSkipNullEnabled(true)
+                .setCollectionsMergeEnabled(false)
+                .setPropertyCondition(Conditions.isNotNull());
+        return modelMapper;
+    }
+
+    @Bean(name = "objectWritter")
+    ObjectWriter objectWriter() {
+        ObjectMapper mapper = new ObjectMapper().registerModule(new JavaTimeModule());
+        return mapper.writer().withDefaultPrettyPrinter();
+    }
+    @Bean
+    public OpenAPI springDocOpenApi() {
+        return new OpenAPI()
+                .info(springDocapiInfo())
+                .security(List.of(new SecurityRequirement()));
+
+    }
+
+    Info springDocapiInfo() {
+        return new Info()
+                .title("Title")
+                .description("Descr")
+                .version("0.0.0-SNAPSHOT");
+    }
+}
