+++ src/main/java/com/example/docxconcat/service/CustomMultipartFile.java
@@ -1,0 +1,59 @@
+package com.example.docxconcat.service;
+
+import org.apache.commons.io.FileUtils;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+
+public class CustomMultipartFile implements MultipartFile {
+    private final byte[] input;
+    private final String originalFilename;
+
+    public CustomMultipartFile(byte[] input, String originalFilename) {
+        this.input = input;
+        this.originalFilename = originalFilename;
+    }
+
+    @Override
+    public String getName() {
+        return null;
+    }
+
+    @Override
+    public String getOriginalFilename() {
+        return originalFilename;
+    }
+
+    @Override
+    public String getContentType() {
+        return null;
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return input == null || input.length == 0;
+    }
+
+    @Override
+    public long getSize() {
+        return input.length;
+    }
+
+    @Override
+    public byte[] getBytes() {
+        return input;
+    }
+
+    @Override
+    public InputStream getInputStream() {
+        return new ByteArrayInputStream(input);
+    }
+
+    @Override
+    public void transferTo(File destination) throws IOException, IllegalStateException {
+        FileUtils.writeByteArrayToFile(destination, input);
+    }
+}
