+++ src/main/java/com/example/docxconcat/web/controller/SwaggerController.java
@@ -1,0 +1,24 @@
+package com.example.docxconcat.web.controller;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+/**
+ * @author ogbozoyan
+ * @date 08.07.2023
+ */
+@CrossOrigin(origins = "*", allowedHeaders = "*")
+@RestController
+@RequestMapping("/")
+public class SwaggerController {
+    @GetMapping(value = "/")
+    public void root(HttpServletResponse rsp) throws IOException {
+        rsp.sendRedirect("swagger-ui/index.html");
+    }
+}
+
