+++ src/main/resources/banner.txt
@@ -1,0 +1,8 @@
+,--.   ,--.                           ,------.   ,-----. ,-----.,--.   ,--.
+|   `.'   |,---. ,--.--. ,---.  ,---. |  .-.  \ '  .-.  '  .--./ \  `.'  /
+|  |'.'|  | .-. :|  .--'| .-. || .-. :|  |  \  :|  | |  |  |      .'    \
+|  |   |  \   --.|  |   ' '-' '\   --.|  '--'  /'  '-'  '  '--'\ /  .'.  \
+`--'   `--'`----'`--'   .`-  /  `----'`-------'  `-----' `-----''--'   '--'
+                        `---'
+${application.title} ${application.version}
+Powered by Spring Boot ${spring-boot.version}
