+++ src/test/java/com/example/docxconcat/MergeDOCXTests.java
@@ -1,0 +1,13 @@
+package com.example.docxconcat;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MergeDOCXTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
